tests: KafkaServers is used by integration tests only
