+++ tests/integration/test_master_coordinator.py
@@ -7,7 +7,8 @@ See LICENSE for details
 from contextlib import closing
 from karapace.config import set_config_defaults
 from karapace.master_coordinator import MasterCoordinator
-from tests.utils import get_random_port, KafkaServers, new_random_name, TESTS_PORT_RANGE
+from tests.integration.utils.kafka_server import KafkaServers
+from tests.utils import get_random_port, new_random_name, TESTS_PORT_RANGE
 
 import asyncio
 import pytest
